dashboard/requirements.txt
faicons
shiny
seaborn
pandas
pybaseball
shinywidgets
plotly.express
